Move auth routes to file
